+++ frontend/techpick-extension/src/widgets/TagAutocompleteDialog/TagAutocompleteDialog.css.ts
